chore: Prepare folder structure and configure gitignore
Add a gitkeep file in data/csv and change gitignore to ignore all files
in this folder except this file.

Since raw data relating to the databases and the CSV resulting files
have a substantial size, and since they are not necessary to install
the project as an R package, these files will be stored elsewhere.