data: Correct the assigment of establishment IDs
In the script creating identifiers for establishments according to their
names, the last part of the three-part identifiers was not created as
expected for establishment names considered to correspond to the same
establishment and to be merged.
The identifiers created before that, on which this three-part
identifiers are based, are correct. But the assignment of the new
identifiers is wrong. It relied on the table function, counting the
number of appearances of each unique value, but also reordering the
result according to the natural order of the values.
The issue was that these values were supposed to be numeric values but
were character values, leading to an alphabetic order instead of a
numeric order (1, 1001, 1002... instead of 1, 2, 3...).