data: Add a script to create USIS datasets
Add a script creating USIS datasets.
Add functions to help creating the datasets.
Add packages in Suggests section of DESCRIPTION.
Dependance to R 2.10 was added automatically to DESCRIPTION because of
the use of the package usethis.

USIS datasets have not been added for now to avoid versioning heavy
files about data that are not ready yet.