fix: Correct joins of reference tables
Join of measures with reference tables by using identifiers as factors
did not work in function join_db. This resulted in wrong associations
id-name that do not correspond to associations existing in reference
tables.
In function join_db, factor values have now been converted to character
before using them so that they correspond to the names or rownames of
the other datasets to join, and can be used for correct indexing.
Since NA character values cannot be used for row indexing in matrices,
subsetting is performed in two steps. First, the desired column is
selected. This results in a vector, in which NA character values can be
used for indexing.