data: Make message about invalid tables clearer
Display a text before the validation of keys in supplementary tables to
explain that next instructions may correct some invalid cases.
Add a second validation after theses instructions.
Precede this second validation with a message mentioning the importance
of this validation.
Change the function validate_keys to return a boolean about the
validity.
Use it on the second validation to stop the script.